# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/group": {
      "post": {
        "operationId": "GroupsController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createGroupDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "get": {
        "operationId": "GroupsController_retrieveAllGroup",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Retrieved all groups successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      }
    },
    "/group/member": {
      "post": {
        "operationId": "GroupsController_addUserToGP",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addUserDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User added to group successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "put": {
        "operationId": "GroupsController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/addUserDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteUser",
        "parameters": [
          {
            "name": "groupId",
            "required": false,
            "in": "query",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/deleteMemberDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User(s) deleted successfully."
          },
          "400": {
            "description": "Invalid parameters."
          }
        },
        "tags": [
          "FGroup"
        ]
      }
    },
    "/group/{id}": {
      "post": {
        "operationId": "GroupsController_createGroupByID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createGroupDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Group created successfully by ID."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "get": {
        "operationId": "GroupsController_retrieveGpById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved group by ID successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "put": {
        "operationId": "GroupsController_updateGP",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/createGroupDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Group updated successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      },
      "delete": {
        "operationId": "GroupsController_deleteGP",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Group ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Group deleted successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      }
    },
    "/group/member/{id}": {
      "delete": {
        "operationId": "GroupsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Member ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Member deleted successfully."
          },
          "400": {
            "description": "Invalid input."
          }
        },
        "tags": [
          "FGroup"
        ]
      }
    },
...
  },
  "info": {
    "title": "OIDC Protocol setup",
    "description": "This is the project to implement OIDC protocol in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "createGroupDTO": {
        "type": "object",
        "properties": {
          "roleIDs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "roleIDs"
        ]
      },
      "addUserDTO": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "members"
        ]
      },
      "deleteMemberDTO": {
        "type": "object",
        "properties": {
          "members": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "members"
        ]
      },
...
    }
  }
}